requirements.txt
# =============================================================================
# Salt Exploration Analysis System
# Requirements File
# =============================================================================

# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Machine learning
scikit-learn>=1.0.0
xgboost>=1.5.0
joblib>=1.1.0

# Image processing
Pillow>=8.0.0
scikit-image>=0.19.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Web application
streamlit>=1.20.0

# AI explanation (Groq LLM API)
groq>=0.4.0

# Progress tracking
tqdm>=4.62.0
